kdt2_TIL/week16/day2/training/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==3.2.18
django-appconf==1.0.5
django-extensions==3.2.1
django-imagekit==4.1.0
executing==1.2.0
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.14.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
